+++ test/test_bsd_ai_packing.c
@@ -50,7 +50,7 @@ int main() {
     0x00, 0x00
   };
   
-  for(int i = 0; i < sizeof(template); i++) {
+  for(size_t i = 0; i < sizeof(template); i++) {
     if(buf[i] != template[i]) {
       dbg_printf("packing mismatch at +0x%x", i);
       hexdump(buf, 0x50);
